[refactor] 메모 삭제 관련 컨트롤러 로직 개선 #93
